chore: move ReceiveSummary under SharedTradeInputFooter directory
